# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Affiliate Aggregator Engine API
  version: 1.0.0
  description: |
    Multi-platform e-commerce product search aggregator with affiliate link cloaking.
    Searches Amazon, Flipkart, Meesho, Croma, and eBay simultaneously.

servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.affiliate-engine.com
    description: Production

paths:
  /api/search:
    get:
      summary: Search products across all platforms
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search query
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, relevance]
            default: price_asc
        - name: merchant
          in: query
          schema:
            type: string
            enum: [amazon, flipkart, meesho, croma, ebay]
      responses:
        '200':
          description: Successful search response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid query
        '429':
          description: Rate limit exceeded

  /api/go/{shortCode}:
    get:
      summary: Resolve cloaked affiliate link (301 redirect)
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 8
      responses:
        '301':
          description: Redirect to raw affiliate URL
        '404':
          description: Short code not found

  /api/stats:
    get:
      summary: Pipeline statistics
      responses:
        '200':
          description: Stats response
          content:
            application/json:
              schema:
                type: object
                properties:
                  cache:
                    $ref: '#/components/schemas/CacheStats'
                  redirects:
                    type: object
                    properties:
                      total:
                        type: integer
                      cloaked_links:
                        type: integer
                  aggregations:
                    type: object
                    properties:
                      total:
                        type: integer

  /api/health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy

components:
  schemas:
    ProductResult:
      type: object
      properties:
        product_id:
          type: string
        title:
          type: string
        normalized_title:
          type: string
        description:
          type: string
        price:
          $ref: '#/components/schemas/Price'
        merchant:
          $ref: '#/components/schemas/Merchant'
        images:
          $ref: '#/components/schemas/Images'
        identifiers:
          $ref: '#/components/schemas/Identifiers'
        match:
          $ref: '#/components/schemas/MatchInfo'
        affiliate:
          $ref: '#/components/schemas/AffiliateInfo'
        metadata:
          $ref: '#/components/schemas/Metadata'

    Price:
      type: object
      properties:
        current_price:
          type: number
        original_price:
          type: number
          nullable: true
        currency_code:
          type: string
          description: ISO 4217 currency code

    Merchant:
      type: object
      properties:
        platform:
          type: string
          enum: [amazon, flipkart, meesho, croma, ebay]
        merchant_name:
          type: string
        merchant_logo_url:
          type: string

    Images:
      type: object
      properties:
        thumbnail_url:
          type: string
        full_url:
          type: array
          items:
            type: string

    Identifiers:
      type: object
      properties:
        model_number:
          type: string
          nullable: true
        upc:
          type: string
          nullable: true
        ean:
          type: string
          nullable: true
        asin:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true

    MatchInfo:
      type: object
      properties:
        matched:
          type: boolean
        match_method:
          type: string
          enum: [exact_code, fuzzy_tokens, none]
        match_confidence:
          type: number
          minimum: 0
          maximum: 1
        match_group_id:
          type: string

    AffiliateInfo:
      type: object
      properties:
        cloaked_url:
          type: string
        raw_url:
          type: string
        merchant_tag:
          type: string

    Metadata:
      type: object
      properties:
        fetched_at:
          type: string
          format: date-time
        cache_hit:
          type: boolean
        stale:
          type: boolean

    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean

    SearchResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductResult'
        pagination:
          $ref: '#/components/schemas/Pagination'

    CacheStats:
      type: object
      properties:
        total:
          type: integer
        stale:
          type: integer
        avgHits:
          type: integer